+++ src/view/board.js
@@ -1,0 +1,7 @@
+export const createBoardTemplate = () => {
+  return (
+    `<section class="board container">
+      <div class="board__tasks"></div>
+    </section>`
+  );
+};
